requirements.txt
fastapi
uvicorn
google-genai
faiss-cpu
numpy
python-dotenv
pydantic
groq
sentence-transformers
requests
python-multipart

